[rust] Move all github commands to same module
